% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwapor_favorites.R
\name{get_favorites_path}
\alias{get_favorites_path}
\title{Get Path to Rwanda WaPOR Favorites File}
\usage{
get_favorites_path()
}
\value{
Character path to the favorites JSON file
}
\description{
Get Path to Rwanda WaPOR Favorites File
}
\keyword{internal}
