% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{guess_l3_region}
\alias{guess_l3_region}
\title{Guess L3 Region from Spatial Intersection}
\usage{
guess_l3_region(variable, reg_info, period)
}
\arguments{
\item{variable}{Character. The WaPOR variable name (e.g., L3-AETI-D)}

\item{reg_info}{List from parse_region()}

\item{period}{Date period vector}
}
\value{
Character vector of intersecting L3 regions, or NULL
}
\description{
Guess L3 Region from Spatial Intersection
}
\keyword{internal}
