% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomaly.R
\name{linear_trend}
\alias{linear_trend}
\title{Calculate per-pixel temporal linear trend}
\usage{
linear_trend(stack, times = NULL)
}
\arguments{
\item{stack}{Multi-layer SpatRaster ordered in time.}

\item{times}{Optional numeric time values, defaulting to layer sequence.}
}
\value{
A list containing SpatRaster layers \code{slope}, \code{intercept}, and \code{r2}.
}
\description{
Calculate per-pixel temporal linear trend
}
