% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwapor_favorites.R
\name{rwapor_add_favorite}
\alias{rwapor_add_favorite}
\title{Add a Path to Favorites}
\usage{
rwapor_add_favorite(path, type = c("directory", "file"))
}
\arguments{
\item{path}{Character. Path to a directory or file.}

\item{type}{Character. One of "directory" or "file".}
}
\value{
Logical. TRUE if successful.
}
\description{
Add a Path to Favorites
}
