% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_indicators.R
\name{rwapor_aggregate_precip_monthly}
\alias{rwapor_aggregate_precip_monthly}
\title{Aggregate Precipitation to Monthly Totals}
\usage{
rwapor_aggregate_precip_monthly(precip_ts)
}
\arguments{
\item{precip_ts}{data.frame with columns: date, value (daily or dekadal precip).}
}
\value{
A data.frame with columns: year, month, p_monthly_mm.
}
\description{
Aggregate Precipitation to Monthly Totals
}
