% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_indicators.R
\name{rwapor_calc_yield_npp}
\alias{rwapor_calc_yield_npp}
\title{Calculate Crop Yield from NPP}
\usage{
rwapor_calc_yield_npp(npp_gc_m2, MC, fc, AOT, HI)
}
\arguments{
\item{npp_gc_m2}{Numeric. Seasonal sum of NPP in gC/m\eqn{^2}.}

\item{MC}{Numeric. Moisture content (fraction, 0-1).}

\item{fc}{Numeric. Light use efficiency correction factor.}

\item{AOT}{Numeric. Above-ground over total biomass production ratio.}

\item{HI}{Numeric. Harvest index (fraction, 0-1).}
}
\value{
Numeric. Crop yield in t/ha.
}
\description{
Calculates crop yield from NPP using the FAO WaPOR biomass model:
\enumerate{
\item \code{DMP = NPP * 22.222} (gC/m\eqn{^2} to kgDM/ha)
\item \code{AGBM = (AOT * fc * (DMP / (1 - MC))) / 1000} (t/ha)
\item \code{Yield = HI * AGBM} (t/ha)
}
}
\seealso{
\code{\link{rwapor_convert_npp_to_tbp}}, \code{\link{rwapor_calc_cwp}}
}
