% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwapor_favorites.R
\name{rwapor_is_favorite}
\alias{rwapor_is_favorite}
\title{Check if a Path is Favorited}
\usage{
rwapor_is_favorite(path)
}
\arguments{
\item{path}{Character. Path to check.}
}
\value{
Logical. TRUE if favorited.
}
\description{
Check if a Path is Favorited
}
