% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_defaults.R
\name{rwapor_list_crops}
\alias{rwapor_list_crops}
\title{List Available Crop Profiles}
\usage{
rwapor_list_crops()
}
\value{
Character vector of crop names.
}
\description{
Returns the names of all saved crop profiles in the FAO defaults dataset.
}
\examples{
rwapor_list_crops()
}
\seealso{
\code{\link{FAO_CROP_DEFAULTS}}, \code{\link{rwapor_get_crop_defaults}}
}
