% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_indicators.R
\name{rwapor_merge_analysis_timeseries}
\alias{rwapor_merge_analysis_timeseries}
\title{Merge Analysis Time Series on Common Time Axis}
\usage{
rwapor_merge_analysis_timeseries(aeti_ts, ret_ts, precip_ts)
}
\arguments{
\item{aeti_ts}{data.frame from wapor_ts().}

\item{ret_ts}{data.frame from wapor_ts().}

\item{precip_ts}{data.frame from wapor_ts().}
}
\value{
A merged data.frame.
}
\description{
Joins AETI, RET, and precipitation data.frames on their start_date column.
}
