% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwapor_favorites.R
\name{rwapor_remove_favorite}
\alias{rwapor_remove_favorite}
\title{Remove a Path from Favorites}
\usage{
rwapor_remove_favorite(path)
}
\arguments{
\item{path}{Character. Path to remove.}
}
\value{
Logical. TRUE if successful.
}
\description{
Remove a Path from Favorites
}
