% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_utils.R
\name{wapor_build_season_profile_table}
\alias{wapor_build_season_profile_table}
\title{Build a table of unique season profiles (start/end combinations)}
\usage{
wapor_build_season_profile_table(
  crop_mask,
  start_raster,
  end_raster,
  class_values
)
}
\arguments{
\item{crop_mask}{SpatRaster.}

\item{start_raster}{SpatRaster.}

\item{end_raster}{SpatRaster.}

\item{class_values}{Integer vector.}
}
\value{
data.frame of profiles.
}
\description{
Build a table of unique season profiles (start/end combinations)
}
\keyword{internal}
