% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_indicators.R
\name{wapor_calc_adequacy_etc}
\alias{wapor_calc_adequacy_etc}
\title{Compute ETc-Based Adequacy}
\usage{
wapor_calc_adequacy_etc(aeti_seasonal, etc_seasonal)
}
\arguments{
\item{aeti_seasonal}{SpatRaster or numeric. Seasonal AETI.}

\item{etc_seasonal}{SpatRaster or numeric. Seasonal ETc.}
}
\value{
SpatRaster or numeric of adequacy ratio.
}
\description{
Adequacy_ETc = Seasonal_AETI / Seasonal_ETc.
Measures whether water consumption meets crop water requirements. Values near 1.0
represent full satisfaction, values below 1.0 indicate water deficit or stress,
and values above 1.0 indicate water application in excess of standard crop demand.
}
\references{
Molden, D. J., & Gates, T. K. (1990). Performance metrics for evaluation of
irrigation-water-delivery systems. Journal of Irrigation and Drainage Engineering, 116(6), 804-823.

Karimi, P., Bastiaanssen, W. G., & Molden, D. (2019). Water accounting plus (WA+) -
a water accounting procedure for complex river basins based on satellite measurements.
Hydrology and Earth System Sciences, 17(7), 2459-2472.
}
