% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{wapor_class_defaults}
\alias{wapor_class_defaults}
\title{Classification schemes and defaults}
\usage{
wapor_class_defaults(scheme = NULL)
}
\arguments{
\item{scheme}{Optional scheme name.}
}
\value{
\code{wapor_class_defaults()} returns a scheme list, or all schemes.
}
\description{
Classification schemes and defaults
}
\examples{
wapor_class_defaults("adequacy")
}
