% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{wapor_class_info}
\alias{wapor_class_info}
\title{Retrieve classification metadata}
\usage{
wapor_class_info(x)
}
\arguments{
\item{x}{A classified factor or SpatRaster.}
}
\value{
The \code{wapor_classes} metadata list, or \code{NULL}.
}
\description{
Retrieve classification metadata
}
\examples{
wapor_class_info(wapor_classify(1:3, breaks = 2:3))
}
