% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{wapor_compare_geom}
\alias{wapor_compare_geom}
\title{Exported Geometry Comparison}
\usage{
wapor_compare_geom(r1, r2)
}
\arguments{
\item{r1}{SpatRaster. First raster to compare.}

\item{r2}{SpatRaster. Second raster to compare.}
}
\value{
Logical.
}
\description{
Exported Geometry Comparison
}
