% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_utils.R
\name{wapor_filter_class_stats}
\alias{wapor_filter_class_stats}
\title{Filter class stats to match active crop parameters}
\usage{
wapor_filter_class_stats(res)
}
\arguments{
\item{res}{Analysis results list.}
}
\value{
data.frame of filtered stats.
}
\description{
Filter class stats to match active crop parameters
}
\keyword{internal}
