% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wapor_monitoring.R
\name{wapor_generate_seasonal_raster}
\alias{wapor_generate_seasonal_raster}
\title{Generate seasonal aggregate raster from DuckDB dekadal blobs}
\usage{
wapor_generate_seasonal_raster(
  con,
  farm_id,
  farm_geom,
  variable,
  start_date,
  end_date
)
}
\arguments{
\item{con}{DuckDB connection}

\item{farm_id}{Farm identifier}

\item{farm_geom}{sf object for the farm}

\item{variable}{WaPOR variable}

\item{start_date}{Seasonal start date}

\item{end_date}{Seasonal end date}
}
\value{
terra SpatRaster aggregate or NULL
}
\description{
Generate seasonal aggregate raster from DuckDB dekadal blobs
}
