% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwapor_favorites.R
\name{wapor_get_favorites}
\alias{wapor_get_favorites}
\title{Load Favorites from Local Storage}
\usage{
wapor_get_favorites()
}
\value{
A data frame of favorites with 'path' and 'type' columns
}
\description{
Load Favorites from Local Storage
}
