% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{wapor_l3_regions_to_df}
\alias{wapor_l3_regions_to_df}
\title{Convert L3_REGIONS list to data.frame}
\usage{
wapor_l3_regions_to_df(reg_list)
}
\description{
Convert L3_REGIONS list to data.frame
}
\keyword{internal}
