% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_engine.R
\name{wapor_load_and_harmonize}
\alias{wapor_load_and_harmonize}
\title{Internal helper to load and harmonize (internal usage)}
\usage{
wapor_load_and_harmonize(paths, template, reg_info, method = "near")
}
\description{
Internal helper to load and harmonize (internal usage)
}
\keyword{internal}
