% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_utils.R
\name{wapor_masked_global_mean}
\alias{wapor_masked_global_mean}
\title{Compute masked global mean}
\usage{
wapor_masked_global_mean(r, mask_rast = NULL)
}
\arguments{
\item{r}{SpatRaster.}

\item{mask_rast}{SpatRaster. Optional mask.}
}
\value{
Numeric mean.
}
\description{
Compute masked global mean
}
\keyword{internal}
