% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators_math.R
\name{wapor_math_adequacy_etc}
\alias{wapor_math_adequacy_etc}
\title{ETc-based adequacy: AETI / ETc}
\usage{
wapor_math_adequacy_etc(aeti_seasonal, etc_seasonal)
}
\description{
ETc-based adequacy: AETI / ETc
}
\keyword{internal}
