% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators_math.R
\name{wapor_math_area_weighted_mean}
\alias{wapor_math_area_weighted_mean}
\title{Area-weighted mean of a numeric matrix or vector}
\usage{
wapor_math_area_weighted_mean(values, area = NULL)
}
\description{
Area-weighted mean of a numeric matrix or vector
}
\keyword{internal}
