% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators_math.R
\name{wapor_math_beneficial_fraction}
\alias{wapor_math_beneficial_fraction}
\title{Beneficial fraction: T / AETI}
\usage{
wapor_math_beneficial_fraction(t_seasonal, aeti_seasonal)
}
\description{
Beneficial fraction: T / AETI
}
\keyword{internal}
