% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators_math.R
\name{wapor_math_blue_water}
\alias{wapor_math_blue_water}
\title{Blue water = max(0, AETI - Peff)}
\usage{
wapor_math_blue_water(aeti_seasonal, peff_seasonal)
}
\description{
Blue water = max(0, AETI - Peff)
}
\keyword{internal}
