% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators_math.R
\name{wapor_math_bwp}
\alias{wapor_math_bwp}
\title{Biomass water productivity in kg/m3}
\usage{
wapor_math_bwp(biomass_value, aeti_mm, biomass_unit = "kg/ha")
}
\description{
Biomass water productivity in kg/m3
}
\keyword{internal}
