% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators_math.R
\name{wapor_math_cwp}
\alias{wapor_math_cwp}
\title{Crop water productivity in kg/m3}
\usage{
wapor_math_cwp(yield_value, aeti_mm, yield_unit = "kg/ha")
}
\description{
Crop water productivity in kg/m3
}
\keyword{internal}
