% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators_math.R
\name{wapor_math_green_water}
\alias{wapor_math_green_water}
\title{Green water = min(AETI, Peff)}
\usage{
wapor_math_green_water(aeti_seasonal, peff_seasonal)
}
\description{
Green water = min(AETI, Peff)
}
\keyword{internal}
