% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators_math.R
\name{wapor_math_npp_to_biomass}
\alias{wapor_math_npp_to_biomass}
\title{NPP (gC/m2) to biomass (kg DM/ha)}
\usage{
wapor_math_npp_to_biomass(npp_gc_m2)
}
\description{
NPP (gC/m2) to biomass (kg DM/ha)
}
\keyword{internal}
