% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators_math.R
\name{wapor_math_peff_usda}
\alias{wapor_math_peff_usda}
\title{USDA-SCS effective precipitation from monthly totals (mm)}
\usage{
wapor_math_peff_usda(p_monthly)
}
\description{
USDA-SCS effective precipitation from monthly totals (mm)
}
\keyword{internal}
