% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators_math.R
\name{wapor_math_yield_from_npp}
\alias{wapor_math_yield_from_npp}
\title{Crop yield (t/ha) from seasonal NPP}
\usage{
wapor_math_yield_from_npp(npp_gc_m2, mc, fc, aot, hi)
}
\description{
Crop yield (t/ha) from seasonal NPP
}
\keyword{internal}
