% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators_math.R
\name{wapor_math_zonal_mean_by_class}
\alias{wapor_math_zonal_mean_by_class}
\title{Mean of values within each crop-mask class}
\usage{
wapor_math_zonal_mean_by_class(values, crop_mask, area = NULL)
}
\description{
Mean of values within each crop-mask class
}
\keyword{internal}
