% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wapor_monitoring.R
\name{wapor_raster_from_blob}
\alias{wapor_raster_from_blob}
\title{Extract Raster from DuckDB BLOB}
\usage{
wapor_raster_from_blob(blob_data)
}
\arguments{
\item{blob_data}{Raw vector containing raster BLOB}
}
\value{
terra SpatRaster object
}
\description{
Extract Raster from DuckDB BLOB
}
