% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_indicators.R
\name{wapor_relative_et_stack}
\alias{wapor_relative_et_stack}
\title{Build a stack of monthly relative ET (AETI / ETc)}
\usage{
wapor_relative_et_stack(result)
}
\arguments{
\item{result}{A seasonal-analysis result with \code{monthly_aeti} and
\code{monthly_etc}, or a list with those names.}
}
\value{
A SpatRaster of AETI_t / ETc_t.
}
\description{
Build a stack of monthly relative ET (AETI / ETc)
}
