% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_indicators.R
\name{rwapor_calc_yield_npp}
\alias{rwapor_calc_yield_npp}
\title{Calculate Crop Yield from NPP}
\usage{
rwapor_calc_yield_npp(npp_gc_m2, MC, fc, AOT, HI)
}
\arguments{
\item{npp_gc_m2}{Numeric. Seasonal sum of NPP in gC/m2.}

\item{MC}{Numeric. Moisture content (0-1).}

\item{fc}{Numeric. Light use efficiency correction factor.}

\item{AOT}{Numeric. Above ground over total biomass production ratio.}

\item{HI}{Numeric. Harvest index.}
}
\value{
Numeric. Crop yield in t/ha.
}
\description{
Implementation of the provided yield formula based on NPP:
AGBM = (AOT * fc * (NPP * 22.222 / (1 - MC))) / 1000
CropYield = HI * AGBM
}
