% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_indicators.R
\name{wapor_calc_adequacy_etc}
\alias{wapor_calc_adequacy_etc}
\title{Compute ETc-Based Adequacy}
\usage{
wapor_calc_adequacy_etc(aeti_seasonal, etc_seasonal)
}
\arguments{
\item{aeti_seasonal}{SpatRaster or numeric. Seasonal AETI.}

\item{etc_seasonal}{SpatRaster or numeric. Seasonal ETc.}
}
\value{
SpatRaster or numeric of adequacy ratio.
}
\description{
Adequacy_ETc = Seasonal_AETI / Seasonal_ETc
}
