% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_indicators.R
\name{wapor_calc_beneficial_fraction}
\alias{wapor_calc_beneficial_fraction}
\title{Compute Beneficial Fraction}
\usage{
wapor_calc_beneficial_fraction(t_seasonal, aeti_seasonal)
}
\arguments{
\item{t_seasonal}{SpatRaster or numeric. Seasonal Transpiration (mm).}

\item{aeti_seasonal}{SpatRaster or numeric. Seasonal AETI (mm).}
}
\value{
SpatRaster or numeric of beneficial fraction (0-1).
}
\description{
Beneficial Fraction = Transpiration (T) / Actual Evapotranspiration (AETI)
}
