% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_indicators.R
\name{wapor_calc_etc}
\alias{wapor_calc_etc}
\title{Compute Dekadal ETc}
\usage{
wapor_calc_etc(ret_dekad, kc_dekad)
}
\arguments{
\item{ret_dekad}{SpatRaster. Dekadal RET layers.}

\item{kc_dekad}{SpatRaster or numeric vector. Dekadal Kc values.
If a numeric vector, each value is applied uniformly to the
corresponding layer.}
}
\value{
A SpatRaster of dekadal ETc.
}
\description{
Multiplies dekadal RET by dekadal Kc to produce dekadal ETc.
}
\keyword{internal}
