% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_indicators.R
\name{wapor_convert_npp_tbp}
\alias{wapor_convert_npp_tbp}
\title{Convert NPP to Total Biomass Production (TBP)}
\usage{
wapor_convert_npp_tbp(npp_gc_m2)
}
\arguments{
\item{npp_gc_m2}{Numeric. Seasonal sum of NPP in gC/m2.}
}
\value{
Numeric. TBP in kgDM/ha.
}
\description{
Converts seasonal NPP (gC/m2) to TBP (kgDM/ha) using the
factor 22.222.
}
