% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wapor_monitoring.R
\name{wapor_init_monitoring_db}
\alias{wapor_init_monitoring_db}
\title{Initialize a DuckDB database for farm monitoring}
\usage{
wapor_init_monitoring_db(con)
}
\arguments{
\item{con}{DuckDB connection object.}
}
\value{
The connection object (invisibly).
}
\description{
Initialize a DuckDB database for farm monitoring
}
