% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wapor_safe_project}
\alias{wapor_safe_project}
\title{Safe Terra Projection Wrapper}
\usage{
wapor_safe_project(x, y)
}
\arguments{
\item{x}{SpatVector or SpatRaster}

\item{y}{target CRS}
}
\value{
SpatVector or SpatRaster
}
\description{
Eliminates PROJ database collisions on some Windows environments
when projecting between WGS84 and UTM.
}
